+++ .buildkite/commands/restore-cache.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -e
+
+echo "--- 💾 Restore Cache"
+restore_gradle_dependency_cache || true