+++ .buildkite/commands/save-cache.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+set -euo pipefail
+
+echo "--- 📦 Download Dependencies"
+./gradlew downloadDependencies
+echo ""
+
+echo "--- 💾 Save Cache"
+save_gradle_dependency_cache