+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarksPage.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -70,6 +71,7 @@ fun BookmarksPage(
     onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     bottomInset: Dp,
+    isDarkTheme: Boolean,
 ) {
     val context = LocalContext.current
     val state by bookmarksViewModel.uiState.collectAsStateWithLifecycle()
@@ -90,6 +92,7 @@ fun BookmarksPage(
         bottomInset = bottomInset,
         onClearSearchTapped = onClearSearchTapped,
         onSearchBarClearButtonTapped = onSearchBarClearButtonTapped,
+        isDarkTheme = isDarkTheme,
     )
     LaunchedEffect(episodeUuid) {
         bookmarksViewModel.loadBookmarks(
@@ -139,6 +142,7 @@ private fun Content(
     onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     bottomInset: Dp,
+    isDarkTheme: Boolean,
 ) {
     Box(
         modifier = Modifier
@@ -157,6 +161,7 @@ private fun Content(
                 bottomInset = bottomInset,
                 onClearSearchTapped = onClearSearchTapped,
                 onSearchBarClearButtonTapped = onSearchBarClearButtonTapped,
+                isDarkTheme = isDarkTheme,
             )
 
             is UiState.Empty -> NoBookmarksView(
@@ -190,6 +195,7 @@ private fun BookmarksView(
     onClearSearchTapped: () -> Unit,
     onSearchBarClearButtonTapped: () -> Unit,
     bottomInset: Dp,
+    isDarkTheme: Boolean,
 ) {
     val focusRequester = remember { FocusRequester() }
     LazyColumn(
@@ -208,6 +214,7 @@ private fun BookmarksView(
                         .fillMaxWidth()
                         .padding(horizontal = 16.dp)
                         .padding(top = 16.dp)
+                        .height(32.dp)
                         .focusRequester(focusRequester),
                 )
             }
@@ -266,6 +273,7 @@ private fun BookmarksView(
                 showIcon = state.showIcon,
                 useEpisodeArtwork = state.useEpisodeArtwork,
                 showEpisodeTitle = state.showEpisodeTitle,
+                isDarkTheme = isDarkTheme,
             )
         }
     }
@@ -315,6 +323,7 @@ private fun BookmarksPreview(
             onClearSearchTapped = {},
             onSearchBarClearButtonTapped = {},
             bottomInset = 0.dp,
+            isDarkTheme = false,
         )
     }
 }