+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/share/ShareListCreateTitlePage.kt
@@ -97,7 +97,7 @@ private fun ShareListCreateTitleContent(
                 onDescriptionChange = onDescriptionChange,
                 onDoneClick = onDoneClick,
                 focusRequester = focusRequester,
-                modifier = Modifier.padding(horizontal = 16.dp),
+                modifier = Modifier.padding(start = 16.dp, end = 16.dp, top = 8.dp),
             )
         }
         LazyVerticalGrid(