+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/CancelConfirmationFragment.kt
@@ -32,7 +32,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.fragment.app.viewModels
-import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.bars.BottomSheetAppBar
 import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationButton
@@ -41,6 +40,7 @@ import au.com.shiftyjelly.pocketcasts.compose.buttons.RowOutlinedButton
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH20
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH40
 import au.com.shiftyjelly.pocketcasts.compose.components.rememberViewInteropNestedScrollConnection
+import au.com.shiftyjelly.pocketcasts.compose.extensions.contentWithoutConsumedInsets
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.compose.theme
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
@@ -68,7 +68,7 @@ class CancelConfirmationFragment : BaseDialogFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ) = content {
+    ) = contentWithoutConsumedInsets {
         AppThemeWithBackground(theme.activeTheme) {
             CancelConfirmationPage(
                 rows = getRows(),
@@ -234,7 +234,7 @@ class CancelConfirmationFragment : BaseDialogFragment() {
             ),
         ),
         Row(
-            iconResId = R.drawable.ic_locked_large,
+            iconResId = IR.drawable.ic_locked_large,
             text = stringResource(LR.string.profile_cancel_confirm_item_plus),
         ),
         Row(
@@ -242,7 +242,7 @@ class CancelConfirmationFragment : BaseDialogFragment() {
             text = stringResource(LR.string.profile_cancel_confirm_item_folders),
         ),
         Row(
-            iconResId = R.drawable.ic_upload___remove_from_cloud___menu,
+            iconResId = IR.drawable.ic_remove_from_cloud,
             text = stringResource(LR.string.profile_cancel_confirm_item_uploads),
         ),
         Row(
@@ -265,7 +265,7 @@ class CancelConfirmationFragment : BaseDialogFragment() {
             CancelConfirmationPage(
                 rows = listOf(
                     Row(
-                        iconResId = R.drawable.ic_locked_large,
+                        iconResId = IR.drawable.ic_locked_large,
                         text = stringResource(LR.string.profile_cancel_confirm_item_plus),
                     ),
                     Row(