+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/winback/WinbackOfferPage.kt
@@ -289,7 +289,6 @@ private fun WinbackOfferPagePreview(
 ) {
     AppThemeWithBackground(
         themeType = theme,
-        backgroundColor = { MaterialTheme.theme.colors.primaryUi04 },
     ) {
         WinbackOfferPage(
             onClaimOffer = {},