+++ modules/features/reimagine/src/main/kotlin/au/com/shiftyjelly/pocketcasts/reimagine/podcast/SharePodcastFragment.kt
@@ -13,12 +13,13 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.core.os.BundleCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.viewModels
-import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
+import au.com.shiftyjelly.pocketcasts.compose.extensions.contentWithoutConsumedInsets
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.reimagine.ui.ShareColors
 import au.com.shiftyjelly.pocketcasts.reimagine.ui.rememberBackgroundAssetControler
 import au.com.shiftyjelly.pocketcasts.sharing.SocialPlatform
+import au.com.shiftyjelly.pocketcasts.ui.helper.StatusBarIconColor
 import au.com.shiftyjelly.pocketcasts.utils.parceler.ColorParceler
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
@@ -46,6 +47,8 @@ class SharePodcastFragment : BaseDialogFragment() {
 
     @Inject internal lateinit var shareListenerFactory: SharePodcastListener.Factory
 
+    override val statusBarIconColor = StatusBarIconColor.Light
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         if (savedInstanceState == null) {
@@ -57,7 +60,7 @@ class SharePodcastFragment : BaseDialogFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ) = content {
+    ) = contentWithoutConsumedInsets {
         val platforms = remember { SocialPlatform.getAvailablePlatforms(requireContext()) }
         val assetController = rememberBackgroundAssetControler(shareColors)
         val listener = remember { shareListenerFactory.create(this@SharePodcastFragment, assetController, args.source) }
@@ -75,10 +78,7 @@ class SharePodcastFragment : BaseDialogFragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        setDialogTint(
-            statusBar = shareColors.navigationBar.toArgb(),
-            navigationBar = shareColors.background.toArgb(),
-        )
+        setDialogTint(color = shareColors.background.toArgb())
     }
 
     @Parcelize