+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/about/AboutFragment.kt
@@ -44,12 +44,12 @@ import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.fragment.app.Fragment
-import androidx.fragment.compose.content
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTracker
@@ -58,6 +58,7 @@ import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
 import au.com.shiftyjelly.pocketcasts.compose.components.HorizontalDivider
+import au.com.shiftyjelly.pocketcasts.compose.extensions.contentWithoutConsumedInsets
 import au.com.shiftyjelly.pocketcasts.compose.theme
 import au.com.shiftyjelly.pocketcasts.localization.BuildConfig
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
@@ -73,6 +74,7 @@ import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import timber.log.Timber
+import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 import au.com.shiftyjelly.pocketcasts.ui.R as UR
 
@@ -87,7 +89,7 @@ class AboutFragment : BaseFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ) = content {
+    ) = contentWithoutConsumedInsets {
         val bottomInset = settings.bottomInset.collectAsStateWithLifecycle(initialValue = 0)
         CallOnce {
             analyticsTracker.track(AnalyticsEvent.SETTINGS_ABOUT_SHOWN)
@@ -113,8 +115,8 @@ class AboutFragment : BaseFragment() {
                 onInstagramTapped = {
                     analyticsTracker.track(AnalyticsEvent.SETTINGS_ABOUT_INSTAGRAM_TAPPED)
                 },
-                onTwitterTapped = {
-                    analyticsTracker.track(AnalyticsEvent.SETTINGS_ABOUT_TWITTER_TAPPED)
+                onXTapped = {
+                    analyticsTracker.track(AnalyticsEvent.SETTINGS_ABOUT_X_TAPPED)
                 },
                 onAutomatticFamilyTapped = {
                     analyticsTracker.track(AnalyticsEvent.SETTINGS_ABOUT_AUTOMATTIC_FAMILY_TAPPED)
@@ -139,21 +141,12 @@ class AboutFragment : BaseFragment() {
 
 private val icons = listOf(
     AppIcon(
-        image = UR.attr.about_logo_wordpress,
-        text = LR.string.settings_about_wordpress,
-        url = "https://wordpress.com/",
-        color = UR.attr.about_logo_wordpress_color,
-        rotate = (-45..45).random(),
-        x = 0.0,
-        y = 23.20,
-    ),
-    AppIcon(
         image = UR.attr.about_logo_jetpack,
         text = LR.string.settings_about_jetpack,
         url = "https://jetpack.com/",
         color = UR.attr.about_logo_jetpack_color,
         rotate = (-45..45).random(),
-        x = 6.17,
+        x = 0.0,
         y = 0.0,
     ),
     AppIcon(
@@ -162,8 +155,8 @@ private val icons = listOf(
         url = "https://dayoneapp.com/",
         color = UR.attr.about_logo_dayone_color,
         rotate = (-45..45).random(),
-        x = 3.83,
-        y = 7.40,
+        x = 6.70,
+        y = 9.40,
     ),
     AppIcon(
         image = UR.attr.about_logo_pocketcasts,
@@ -171,7 +164,7 @@ private val icons = listOf(
         url = "https://www.pocketcasts.com/",
         color = UR.attr.about_logo_pocketcasts_color,
         rotate = 0,
-        x = 2.77,
+        x = 3.37,
         y = 0.0,
     ),
     AppIcon(
@@ -180,7 +173,7 @@ private val icons = listOf(
         url = "https://woocommerce.com/",
         color = UR.attr.about_logo_woo_color,
         rotate = (-45..45).random(),
-        x = 1.94,
+        x = 2.13,
         y = 17.28,
     ),
     AppIcon(
@@ -189,7 +182,7 @@ private val icons = listOf(
         url = "https://simplenote.com/",
         color = UR.attr.about_logo_simplenote_color,
         rotate = (-45..45).random(),
-        x = 1.49,
+        x = 1.56,
         y = 0.0,
     ),
     AppIcon(
@@ -198,8 +191,8 @@ private val icons = listOf(
         url = "https://tumblr.com/",
         color = UR.attr.about_logo_tumblr_color,
         rotate = (-45..45).random(),
-        x = 1.205,
-        y = 19.9,
+        x = 1.225,
+        y = 20.0,
     ),
 )
 
@@ -210,7 +203,7 @@ private fun AboutPage(
     onShareWithFriendsTapped: () -> Unit,
     onWebsiteTapped: () -> Unit,
     onInstagramTapped: () -> Unit,
-    onTwitterTapped: () -> Unit,
+    onXTapped: () -> Unit,
     onAutomatticFamilyTapped: () -> Unit = {},
     onWorkWithUsTapped: () -> Unit = {},
     onTermsOfServiceTapped: () -> Unit = {},
@@ -243,6 +236,7 @@ private fun AboutPage(
                 text = stringResource(LR.string.settings_version, BuildConfig.VERSION_NAME, BuildConfig.VERSION_CODE.toString()),
                 style = MaterialTheme.typography.body1,
                 modifier = Modifier.padding(top = 8.dp),
+                textAlign = TextAlign.Center,
                 color = MaterialTheme.theme.colors.primaryText02,
             )
         }
@@ -294,11 +288,11 @@ private fun AboutPage(
         }
         item {
             RowTextButton(
-                text = stringResource(LR.string.settings_about_twitter),
+                text = stringResource(LR.string.settings_about_x),
                 secondaryText = "@pocketcasts",
                 onClick = {
-                    onTwitterTapped()
-                    openUrl("https://twitter.com/pocketcasts", context)
+                    onXTapped()
+                    openUrl("https://x.com/pocketcasts", context)
                 },
             )
         }
@@ -341,8 +335,38 @@ private fun AboutPage(
             }
         }
         item {
-            Spacer(modifier = Modifier.height(16.dp))
+            Spacer(modifier = Modifier.height(24.dp))
         }
+        item {
+            AutomatticLogo(onAutomatticFamilyTapped = onAutomatticFamilyTapped)
+        }
+        item {
+            Spacer(modifier = Modifier.height(24.dp))
+        }
+    }
+}
+
+@Composable
+fun AutomatticLogo(
+    modifier: Modifier = Modifier,
+    onAutomatticFamilyTapped: () -> Unit = {},
+) {
+    val context = LocalContext.current
+    Box(
+        modifier = modifier
+            .fillMaxWidth()
+            .clickable {
+                onAutomatticFamilyTapped()
+                openUrl("https://automattic.com", context)
+            }
+            .padding(16.dp),
+        contentAlignment = Alignment.Center,
+    ) {
+        Image(
+            painter = painterResource(IR.drawable.about_logo_automattic),
+            contentDescription = stringResource(LR.string.settings_app_icon),
+            colorFilter = ColorFilter.tint(MaterialTheme.theme.colors.primaryText01),
+        )
     }
 }
 
@@ -362,7 +386,7 @@ fun AutomatticFamilyRow(
                 onAutomatticFamilyTapped()
                 openUrl("https://automattic.com", context)
             }
-            .height(192.dp)
+            .height(180.dp)
             .fillMaxWidth(),
         contentAlignment = Alignment.TopStart,
     ) {
@@ -373,7 +397,7 @@ fun AutomatticFamilyRow(
             modifier = Modifier.padding(all = 14.dp),
         )
 
-        val circleWidth = appIconViewWidth / 6.0
+        val circleWidth = appIconViewWidth / 5.5
         icons.forEach { icon ->
             AppLogoImage(
                 width = circleWidth.dp,
@@ -384,7 +408,7 @@ fun AutomatticFamilyRow(
                 onClick = { openUrl(icon.url, context) },
                 modifier = Modifier.offset(
                     x = (if (icon.x == 0.0) 0.0 else appIconViewWidth / icon.x).dp,
-                    y = (192 - circleWidth - (if (icon.y == 0.0) 0.0 else appIconViewWidth / icon.y)).dp,
+                    y = (180 - circleWidth - (if (icon.y == 0.0) 0.0 else appIconViewWidth / icon.y)).dp,
                 ),
             )
         }
@@ -496,7 +520,7 @@ private fun AboutPagePreview() {
         onShareWithFriendsTapped = {},
         onWebsiteTapped = {},
         onInstagramTapped = {},
-        onTwitterTapped = {},
+        onXTapped = {},
         onAutomatticFamilyTapped = {},
         onWorkWithUsTapped = {},
     )