+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/helper/MatchParentFrameLayout.kt
@@ -1,0 +1,19 @@
+package au.com.shiftyjelly.pocketcasts.views.helper
+
+import android.content.Context
+import android.util.AttributeSet
+import android.widget.FrameLayout
+
+/**
+ * A FrameLayout that will always measure itself with the height of the parent.
+ * This fixes an issue with the episode bottom sheet not expanding to the full height of the screen.
+ */
+class MatchParentFrameLayout @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0) : FrameLayout(context, attrs, defStyleAttr) {
+    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
+        var newHeightMeasureSpec = heightMeasureSpec
+        if (MeasureSpec.getMode(heightMeasureSpec) == MeasureSpec.AT_MOST) {
+            newHeightMeasureSpec = MeasureSpec.makeMeasureSpec(MeasureSpec.getSize(heightMeasureSpec), MeasureSpec.EXACTLY)
+        }
+        super.onMeasure(widthMeasureSpec, newHeightMeasureSpec)
+    }
+}